requirements.txt
asyncio==3.4.3
websockets==9.1
pyserial==3.5